bot: user and timekeeping
Keep track of the user who ran the command, as well as the start time of
the command.  This allows the bot to notify the user upon completion of
a long running command.
This can later be expanded to do more cool things like showing the user
who started a job, and how long it's been running, in `!jobs`.

Fixes: #16